Add state saving and loading functionality
Expanded the app's capabilities to save and load data of players' character builds to/from a server. Instead of storing data in localStorage, the application now saves and retrieves data to/from a server with Flask routes. This includes not just the selected state, but also the loaded HTML for the ability traits. In addition, 'Clear Build' button has been added to the interface.